consolidate E and FUV in harmonics.py
