+++ utide/_solve.py
@@ -7,7 +7,7 @@ from __future__ import absolute_import, division
 import numpy as np
 import scipy       # This will go away; see FIXME below
 
-from .ut_E import ut_E
+from .harmonics import ut_E
 from .ut_diagn import ut_diagn
 from .ut_cs2cep import ut_cs2cep
 from .ut_cnstitsel import ut_cnstitsel
